requirements.txt
fire==0.4.0
h5py==3.6.0
hdf5plugin==3.2.0
hydra-core==1.2.0
imageio==2.19.3
omegaconf==2.2.2
opencv-python==4.5.5.64
pandas==1.4.3
pytorch-lightning==1.7.4
PyYAML==6.0
scipy==1.8.1
seaborn==0.11.2
tensorboard==2.10.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
torch==1.12.1+cu113
torchaudio==0.12.1+cu113
torchmetrics==0.9.3
torchvision==0.13.1+cu113
tqdm==4.63.0
prettytable==3.4.1
